Split components from QuizImpl into classes, date fixes
If the user checks the deadline box and doesn't touch the default date, it is saved properly. Before saving to the database,
the chosen deadline is shown as a locally formatted string, not always as a UTC standard string.
